ci: Create major version tag (#148)
This PR will create a tag associated to the major version every time a
docker image is published.

For example, if the current version is `v1.1.0`, it will also create a
`v1` tag.

The purpose of this is that it can be used in CI/local environments
knowing that you will always get the latest version of `v1.x.x`.